Back

+++ Backend/apps/user/views.py
@@ -95,7 +95,7 @@ def get_stuff(request):
         roles = UserRole.objects.filter(user=user)
         role_names = [role.role for role in roles]
         super_admin = False
-        print(role_names)
+      
         if "superadmin" in role_names:
             super_admin = True
         if (not "admin" in role_names) and (not super_admin):
@@ -105,7 +105,7 @@ def get_stuff(request):
         for user in users:
             roles = UserRole.objects.filter(user=user)
             role_names = [role.role for role in roles]
-            if  (not "admin" in role_names or super_admin) and  (not "user" in role_names):
+            if  ((not "admin" in role_names and not "superadmin" in role_names) or super_admin) and  (not "user" in role_names):
                 user_info = {
                         "id_user": user.id_user,
                         "username": user.username,
@@ -130,7 +130,7 @@ def get_users(request):
         for user in users:
             # Получаем информацию о ролях
             roles = UserRole.objects.filter(user=user)
-            role_names = [role.role for role in roles]
+            role_names = list({role.role for role in roles})
             user_info = {
                     "id_user": user.id_user,
                     "username": user.username,
@@ -149,15 +149,18 @@ def get_role(request, id_user):
         user = get_object_or_404(CustomUser, id_user=id_user)
         roles = list(user.roles.values_list("role", flat=True))
         role_list = []
+        manager_process = 0
         for role in roles:
             data = {"role": role}
             if role == "user":
                 user_profile = UserProfile.objects.filter(user_role__user=user).first()
                 data['karma'] = user_profile.karma if user_profile else None
             if role == "manager" :
-                manager_profile = ManagerProfile.objects.filter(user_role__user=user).first()
-                location_id = manager_profile.location_id.id_location if manager_profile and manager_profile.location_id else None
+                manager_profile = ManagerProfile.objects.filter(user_role__user=user)
+                manager_profile = manager_profile[manager_process]
+                location_id = manager_profile.location_id.id_location
                 data['location_id'] = location_id
+                manager_process += 1       
             role_list.append(data)
         data = {
             'id_user': user.id_user,
@@ -178,15 +181,20 @@ def get_user(request, id_user):
         user = get_object_or_404(CustomUser, id_user=id_user)
         roles = list(user.roles.values_list("role", flat=True))
         role_list = []
+        manager_process = 0
         for role in roles:
             data = {"role": role}
+
             if role == "user":
                 user_profile = UserProfile.objects.filter(user_role__user=user).first()
                 data['karma'] = user_profile.karma if user_profile else None
             if role == "manager" :
-                manager_profile = ManagerProfile.objects.filter(user_role__user=user).first()
-                location_id = manager_profile.location_id.id_location if manager_profile and manager_profile.location_id else None
+                manager_profile = ManagerProfile.objects.filter(user_role__user=user)
+                manager_profile = manager_profile[manager_process]
+                location_id = manager_profile.location_id.id_location
                 data['location_id'] = location_id
+                manager_process += 1       
+                
             role_list.append(data)
         data = {
             'id_user': user.id_user,
@@ -229,12 +237,12 @@ def user_create(request):
         login = data.get('login')
         password = data.get('password')
         status = data.get('status', "active")  # По умолчанию "active"
-        location_id = data.get('location_id')
+        location_id = data.get('location_id', [])
         roles = data.get('roles', [])  # roles передается как список строк, например ["user", "manager"]
 
         # Проверка на уникальность username
-        if CustomUser.objects.filter(username=username).exists():
-            return JsonResponse({'error': 'Username already exists'}, status=400)
+        if CustomUser.objects.filter(login=login).exists():
+            return JsonResponse({'error': 'login already exists'}, status=400)
 
         # Создание пользователя
         user = CustomUser(username=username, id_telegram=id_telegram, login=login)
@@ -245,15 +253,18 @@ def user_create(request):
 
         # Привязка ролей
         for role_name in roles:
-            user_role = UserRole.objects.create(user=user, role=role_name)
 
             if role_name == "user":
+                user_role = UserRole.objects.create(user=user, role=role_name)
                 UserProfile.objects.create(user_role=user_role)
 
-            if role_name == "manager":
-                # Получение экземпляра Location
-                location = get_object_or_404(Location, id_location=location_id)
-                ManagerProfile.objects.create(user_role=user_role, location_id=location)
+            elif role_name == "manager":
+                for location_id_obj in location_id:
+                    location = get_object_or_404(Location, id_location=location_id_obj)
+                    user_role = UserRole.objects.create(user=user, role=role_name)
+                    ManagerProfile.objects.create(user_role=user_role, location_id=location)
+            else:
+                user_role = UserRole.objects.create(user=user, role=role_name)
 
         return JsonResponse({
             'id_user': user.id_user,
@@ -282,7 +293,7 @@ def user_update(request, id_user):
         password = data.get('password')
         status = data.get('status')
         location_id = data.get('location_id')
-        roles = data.get('roles', [])
+        roles = data.get('roles')
 
         # Находим пользователя по id_user
         user = get_object_or_404(CustomUser, id_user=id_user)
@@ -302,25 +313,42 @@ def user_update(request, id_user):
         user.save()  # Сохраняем изменения
 
         # Удаление старых ролей
-        UserRole.objects.filter(user=user).delete()
+        if roles is not None:
+            UserRole.objects.filter(user=user).delete()
+            # Привязка новых ролей
+            for role_name in roles:
+                if role_name == "user":
+                    user_role = UserRole.objects.create(user=user, role=role_name)
+                    UserProfile.objects.create(user_role=user_role)
 
-        # Привязка новых ролей
-        for role_name in roles:
-            user_role = UserRole.objects.create(user=user, role=role_name)
+                elif role_name == "manager":
+                    for location_id_obj in location_id:
+                        location = get_object_or_404(Location, id_location=location_id_obj)
+                        user_role = UserRole.objects.create(user=user, role=role_name)
+                        ManagerProfile.objects.create(user_role=user_role, location_id=location)
+                else:
+                    user_role = UserRole.objects.create(user=user, role=role_name)
 
-            if role_name == "user":
-                UserProfile.objects.update_or_create(user_role=user_role)
-
-            if role_name == "manager":
-                location = get_object_or_404(Location, id_location=location_id)
-                ManagerProfile.objects.update_or_create(user_role=user_role, defaults={'location_id': location})
-
+        roles = list(user.roles.values_list("role", flat=True))
+        role_list = []
+        for role in roles:
+            data = {"role": role}
+            if role == "user":
+                user_profile = UserProfile.objects.filter(user_role__user=user).first()
+                data['karma'] = user_profile.karma if user_profile else None
+            if role == "manager" :
+                manager_profile = ManagerProfile.objects.filter(user_role__user=user).first()
+                location_id = manager_profile.location_id.id_location if manager_profile and manager_profile.location_id else None
+                data['location_id'] = location_id
+            role_list.append(data)
+      
+        
         return JsonResponse({
             'id_user': user.id_user,
             'username': user.username,
             'login': user.login if user.login else None,
             'status': user.status,
-            'roles': roles
+            'roles': role_list
         }, status=200)
 
     except json.JSONDecodeError:
